refactor component index location
